pyproject.toml
[project]
name = "catering-backend"
version = "1.0.0"
description = "Catering System Backend for Larvik Kommune"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.12.1",
    "asyncpg==0.29.0",
    "bcrypt==4.3.0",
    "celery==5.3.4",
    "email-validator==2.1.0",
    "faker==20.1.0",
    "fastapi==0.104.1",
    "fuzzywuzzy==0.18.0",
    "python-Levenshtein==0.23.0",
    "greenlet==3.0.3",
    "httpx==0.25.2",
    "openai>=1.0.0",
    "passlib==1.7.4",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.6",
    "redis[hiredis]==5.0.1",
    "reportlab==4.0.8",
    "sqlalchemy==2.0.23",
    "uvicorn[standard]==0.24.0.post1",
]

[project.optional-dependencies]
dev = [
    "black==23.11.0",
    "flake8==6.1.0",
    "isort==5.12.0",
    "mypy==1.7.1",
    "pre-commit==3.5.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-bdd==6.1.1",
    "pytest-cov==4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"